+++ source/YuzuModDownloader/frmMain.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Diagnostics;
+using System.IO;
 using System.Windows.Forms;
 
 namespace YuzuModDownloader
@@ -14,16 +15,13 @@ namespace YuzuModDownloader
         private void frmMain_Load(object sender, EventArgs e)
         {
             // check if software is latest version 
-            if (!AppUpdater.IsLatestVersion())
-            {
-                MessageBox.Show("New version of Yuzu Mod Downloader available, please download from https://github.com/amakvana/YuzuModDownloader", this.Text, MessageBoxButtons.OK, MessageBoxIcon.Error);
-                Process.Start("https://github.com/amakvana/YuzuModDownloader");
-                Application.Exit();
-            }
+            CheckAppVersion();
 
             // set UI defaults 
             lblProgress.Text = "";
-            toolTip1.SetToolTip(chkClearModDataLocation, "Check this to delete all existing mods before downloading the latest switch mods");
+            chkDeleteModArchives.Checked = true;
+            toolTip1.SetToolTip(chkClearModDataLocation, "Deletes all existing mods before downloading the latest Yuzu Game Mods");
+            toolTip1.SetToolTip(chkDeleteModArchives, "Deletes all downloaded mod archive files once unpacked");
             toolTip1.SetToolTip(btnDownload, "Download Yuzu Game Mods for current switch games dumped");
         }
 
@@ -50,10 +48,14 @@ namespace YuzuModDownloader
                 totalGames++;
             }
 
-            // enable form controls
+            // tell user mods have been downloaded 
             MessageBox.Show($"Done! Mods downloaded for {totalGames} games.{Environment.NewLine}{Environment.NewLine}" +
                             $"To toggle specific game mods On/Off:{Environment.NewLine}" +
                             $"Run yuzu > Right-Click on a game > Properties > Add-Ons", this.Text, MessageBoxButtons.OK, MessageBoxIcon.Information);
+
+            // cleanup & reset ui 
+            if (chkDeleteModArchives.Checked) modDownloader.DeleteDownloadedModArchives();
+            if (File.Exists(gameTitleIDsXml)) File.Delete(gameTitleIDsXml);
             ToggleControls(true);
         }
 
@@ -83,10 +85,39 @@ namespace YuzuModDownloader
             }
         }
 
+        private void CheckForUpdatesToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            CheckAppVersion(true);
+        }
+
         private void ToggleControls(bool value)
         {
             btnDownload.Enabled = value;
             chkClearModDataLocation.Enabled = value;
+            chkDeleteModArchives.Enabled = value;
+            lblProgress.Text = "";
+            pbarProgress.Value = 0;
+
+        }
+
+        private void CheckAppVersion(bool manualCheck = false)
+        {
+            var currentAppVersion = AppUpdater.CheckVersion();
+            switch (currentAppVersion)
+            {
+                case AppUpdater.CurrentVersion.LatestVersion when manualCheck:
+                    MessageBox.Show("You currently have the latest version of Yuzu Mod Downloader", Text, MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    break;
+                case AppUpdater.CurrentVersion.UpdateAvailable:
+                    MessageBox.Show("New version of Yuzu Mod Downloader available, please download from https://github.com/amakvana/YuzuModDownloader", Text, MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    Process.Start("https://github.com/amakvana/YuzuModDownloader");
+                    break;
+                case AppUpdater.CurrentVersion.NotSupported:
+                    MessageBox.Show("This version of Yuzu Mod Downloader is no longer supported, please download the latest version from https://github.com/amakvana/YuzuModDownloader", Text, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    Process.Start("https://github.com/amakvana/YuzuModDownloader");
+                    Application.Exit();
+                    break;
+            }
         }
     }
 }